DeepRecurrentEncoder

Documentation for DeepRecurrentEncoder.

DeepRecurrentEncoder.DRE

DeepRecurrentEncoder.DRE

DeepRecurrentEncoder.DRE

DeepRecurrentEncoder.add_mask

DeepRecurrentEncoder.add_mask!

DeepRecurrentEncoder.add_stimuli

DeepRecurrentEncoder.add_stimuli!

DeepRecurrentEncoder.compute_loss

DeepRecurrentEncoder.create_optimiser

DeepRecurrentEncoder.generate_designmatrix

DeepRecurrentEncoder.mse

DeepRecurrentEncoder.prepare_data

DeepRecurrentEncoder.r_squared

DeepRecurrentEncoder.train_test_split

StatsAPI.fit

DeepRecurrentEncoder.DRE — Type

DRE{L,E,D}

Defines a Deep Recurrent Encoder (DRE) structure with LSTM, encoder, and decoder components.

Fields

lstm_cell::L: The LSTM cell component of the DRE.

encoder::E: The encoder component of the DRE.

decoder::D: The decoder component of the DRE.

DeepRecurrentEncoder.DRE — Method

Constructs a DRE model with specified input, hidden, and output channels.

Arguments

in_chs::Int: Number of input channels.

hidden_chs::Int: Number of hidden channels (filters in convolution and units in LSTM).

out_chs::Int: Number of output channels.

kernel_size=4: Kernel size for convolutional layers. Defaults to 4.

stride=2: Stride for convolutional layers. Defaults to 2.

Returns

A new DRE model instance.

DeepRecurrentEncoder.DRE — Method

Applies the DRE model to the input data.

Arguments

s::DRE: The DRE model instance.

x::AbstractArray{T,3}: Input data with three dimensions (time, channel, epoch).

ps::NamedTuple: Parameter states of the model.

st::NamedTuple: Training states of the model.

Returns

decoded: The decoded output after applying the model.

st: Updated states of the model components.

DeepRecurrentEncoder.add_mask! — Method

add_mask!(masked, eeg, p)

In-place version of add_mask.

Arguments

masked::AbstractArray{T,3}: Array to store masked EEG data.

eeg::AbstractArray{T,3}: Input EEG data of shape (channels, time, epochs).

p::Float64: Proportion of data to mask.

Returns

AbstractArray: The masked array.

DeepRecurrentEncoder.add_mask — Method

add_mask(eeg, p)

Add a mask to the EEG data.

Arguments

eeg::AbstractArray: The input EEG data of shape (channels, time, epochs).

p::Float64: The proportion of data to mask.

Returns

AbstractArray: A new array containing the EEG data with a mask added.

DeepRecurrentEncoder.add_stimuli! — Method

add_stimuli!(out, in, designmatrix)

In-place version of add_stimuli.

Arguments

out: Output array to store the result.

in: EEG data of shape (time, channels, epochs).

designmatrix: Matrix describing stimuli for each epoch.

Returns

AbstractArray: The output array with stimuli added.

DeepRecurrentEncoder.add_stimuli — Method

add_stimuli(in, designmatrix)

Add stimuli information to EEG data.

Arguments

in::AbstractArray{T,3}: EEG data of shape (time, channels, epochs).

designmatrix::AbstractArray{T,2}: Matrix describing stimuli for each epoch.

Returns

AbstractArray: EEG data with stimuli added.

DeepRecurrentEncoder.compute_loss — Function

compute_loss(x, y, model, ps, st; loss_function=mse)

Compute the loss for the given inputs, model, and loss function.

Arguments

x: Input data.

y: Ground truth values.

model: Function representing the predictive model.

ps: Model parameters.

st: Model state.

loss_function: Loss function to use (default: mse).

Returns

(Float64, AbstractArray, Any): Tuple of loss, predicted values, and updated model state.

DeepRecurrentEncoder.create_optimiser — Function

Creates an optimizer for training the model.

Arguments

ps: Model parameters to optimize.

lr=0.001: Learning rate for the optimizer. Defaults to 0.001.

Returns

A setup optimizer instance using the ADAM algorithm.

DeepRecurrentEncoder.generate_designmatrix — Method

generate_designmatrix(f, evts)

Generate the design matrix for given formula and events.

Arguments

f::FormulaTerm: A formula describing the design.

evts: Events data.

Returns

AbstractArray: The design matrix.

DeepRecurrentEncoder.mse — Method

mse(y_pred, y_true)

Calculate the mean squared error (MSE) between predicted and true values.

Arguments

y_pred::AbstractArray: Predicted values.

y_true::AbstractArray: True values.

Returns

Float64: The computed MSE.

DeepRecurrentEncoder.prepare_data — Method

Prepares the input data for training by adding masks and stimuli.

Arguments

data: Input data array.

designmatrix: The design matrix to integrate into the input data.

mask_percentage=0.3: The percentage of data to mask. Defaults to 0.3.

kwargs...: Additional keyword arguments.

Returns

input_data: Prepared input data with masks and stimuli added.

data: The original data with dimensions potentially permuted.

DeepRecurrentEncoder.r_squared — Method

r_squared(y_pred, y_true)

Compute the coefficient of determination (R²) for the given predictions.

Arguments

y_pred::AbstractArray: Predicted values.

y_true::AbstractArray: True values.

Returns

Float64: The R² score.

DeepRecurrentEncoder.train_test_split — Function

train_test_split(data, evts; at=0.7)

Split data and events into training and testing sets.

Arguments

data::AbstractArray: Input data of shape (time, channels, epochs).

evts::AbstractArray: Events data.

at::Float64: Proportion of data to include in the training set (default: 0.7).

Returns

(AbstractArray, AbstractArray, AbstractArray, AbstractArray): Training data, training events, testing data, and testing events.

StatsAPI.fit — Method

StatsModels.fit(t::Type{DRE}, args...; kwargs...)

This function provides a shorthand for calling the StatsModels.fit method with a default random number generator (MersenneTwister(1)).

# Arguments
- `t::Type{DRE}`: The type of the model to fit.
- `args...`: Additional positional arguments passed to the fitting function.
- `kwargs...`: Named arguments passed to the fitting function.

# Returns
- The result of the fitting process using a default RNG.