DeepRecurrentEncoder

Documentation for DeepRecurrentEncoder.

DeepRecurrentEncoder.DRE

DeepRecurrentEncoder.add_mask

DeepRecurrentEncoder.add_mask!

DeepRecurrentEncoder.add_stimuli

DeepRecurrentEncoder.prepare_data

DeepRecurrentEncoder.DRE — Method

hidden_chs corresponds to the number of filters in the conv layer & the number of hidden units in the LSTM

DeepRecurrentEncoder.add_mask! — Method

returns new array containing eeg with mask or writes the same contents into masked if provided

DeepRecurrentEncoder.add_mask — Method

returns new array containing eeg with mask or writes the same contents into masked if provided

DeepRecurrentEncoder.add_stimuli — Method

eeg is a 3d array of shape (channel, time, epoch)

designmatrix is a Matrix with a row discribing the stimuli for each epoch

DeepRecurrentEncoder.prepare_data — Method

Todo: Can be greatly improved by generating the stimulus matrix once and then masking + add_stimuli :shrug: